+++ tests/test_ap_dataset.py
@@ -6,8 +6,11 @@ from rtapipe.lib.dataset.dataset import APDataset
 class TestAPDataset:
 
     def test_multiple_phlist(self):
+        
+        # make me a fixture
+        conf_file = Path(Path(__file__).resolve().parent).joinpath("conf/agilehost3-prod5-test.yml")
 
-        ds = APDataset.get_dataset("./conf/agilehost3-prod5-test.yml", 101, out_dir="./test_multiple_phlist-out", scaler_type="mm")
+        ds = APDataset.get_dataset(conf_file, 101, out_dir="./test_multiple_phlist-out", scaler_type="mm")
 
         ds.loadData()
 
@@ -19,17 +22,23 @@ class TestAPDataset:
         assert train_x.shape == (1, 10, 4)
         assert val_x.shape == (1, 10, 4)
 
-        ds.plotRandomSample()
+        sample = ds.get_random_train_sample(True)
+        assert sample.shape == (10, 4)
+
+        ds.plotRandomSample(howMany=10)
         
         assert Path("./test_multiple_phlist-out/random_sample_scaled_True.png").exists()
 
-        ds.plotRandomSample(scaled=False)
+        ds.plotRandomSample(howMany=10, scaled=False)
 
         assert Path("./test_multiple_phlist-out/random_sample_scaled_False.png").exists()
 
     def test_single_phlist(self):
 
-        ds = APDataset.get_dataset("./conf/agilehost3-prod5-test.yml", 201, out_dir="./test_single_phlist-out", scaler_type="mm")
+        # make me a fixture
+        conf_file = Path(Path(__file__).resolve().parent).joinpath("conf/agilehost3-prod5-test.yml")
+
+        ds = APDataset.get_dataset(conf_file, 1201, out_dir="./test_single_phlist-out", scaler_type="mm")
 
         ds.loadData()
 
@@ -41,10 +50,10 @@ class TestAPDataset:
         assert train_x.shape == (90, 10, 3)
         assert val_x.shape == (90, 10, 3)
 
-        ds.plotRandomSample()
+        ds.plotRandomSample(howMany=10)
 
         assert Path("./test_single_phlist-out/random_sample_scaled_True.png").exists()
 
-        ds.plotRandomSample(scaled=False)
+        ds.plotRandomSample(howMany=10, scaled=False)
 
         assert Path("./test_single_phlist-out/random_sample_scaled_False.png").exists()