+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 * Spotted error on flux. Set flux_error=0 for now. 
 * Changed the number of energy bin (from 4 to 3).
 * Refactor APDataset to be able to process a single time series and to extract sub-sequences.
+* A bug in the simulation code has been spotted and fixed! 
+* The dataset config file now supports 'single' and 'multiple' dataset types
+* New 'single' datasets has been added to the agilehost3-prod5.yml